Address another low-risk Sonar cleanup batch across core and handlers.
This batch simplifies negated boolean/range checks and removes redundant casts while preserving behavior under spotless formatting and -Pmin test validation.

Made-with: Cursor

+++ jacop-core/src/main/java/org/jacop/constraints/cumulative/Task.java
@@ -66,7 +66,7 @@ class Task {
   }
 
   long energy() {
-    return (long) dur.min() * (long) res.min();
+    return (long) dur.min() * res.min();
   }
 
   boolean exists() {
